% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ilogit}
\alias{ilogit}
\title{Calculate a proportion}
\usage{
ilogit(w)
}
\arguments{
\item{w}{A log-odds}
}
\value{
A proportion
}
\description{
Calculate the proportion of a log-odds
}
\details{
Calculate the proportion of a log-odds. exp(w) / (1 + exp(w))
}

