% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logit}
\alias{logit}
\title{Calculate log-odds}
\usage{
logit(p)
}
\arguments{
\item{p}{a proportion}
}
\value{
the log-odds.
}
\description{
Calculate the log-odds of a proportion
}
\details{
Calculate the log-odds of a proportion log(p / (1 - p))
}

